% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kristynFunctions.R
\name{export_test}
\alias{export_test}
\title{Export Test (and Solutions!)}
\usage{
export_test(list_of_QAs, testfile = NULL, solutionsfile = NULL)
}
\arguments{
\item{list_of_QAs}{a list of character vectors}

\item{testfile}{name of test file (if none given, will be named "test.txt")}

\item{solutionsfile}{name of test file (if none given, will be named "solutions.txt")}
}
\value{
produces the two latex files (as .txt files)
}
\description{
Export Test (and Solutions!)
}
\examples{
QA1 = makeQA_CIproportion()
QA2 = makeQA_Normal()
QA3 = makeQA_EventProbability()
QA4 = makeQA_ExpectedValue()
QA5 = makeQA_HypothesisTest()
QA6 = makeQA_ConditionalProbability()
QAs = list(QA1, QA2, QA3, QA4, QA5, QA6)
export_test(QAs)
}
