% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kristynFunctions.R
\name{import_question}
\alias{import_question}
\title{Import Your Own Question}
\usage{
import_question(file)
}
\arguments{
\item{file}{a .txt file that has questions (preceded by a tab and a *Q*) and answer choices (preceded by a tab and a *A*, with correct answer choice ending in \%ans)}
}
\value{
a vector with first element as question, second element as correct answer, and other elements as other answer choices.
}
\description{
Import Your Own Question
}
\examples{
# for a .txt file, call:
# import_question(file)
}
