% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kristynFunctions.R
\name{makeQA_CIproportion}
\alias{makeQA_CIproportion}
\title{Make a Confidence Interval for Proportion Question}
\usage{
makeQA_CIproportion(n = NULL, numPositive = NULL, C = 0.95,
  population = NULL, individuals = "individuals", question = NULL,
  answer = NULL)
}
\arguments{
\item{n}{sample size}

\item{numPositive}{number of positive samples (number of individuals who gave positive responses)}

\item{C}{confidence level (default is 0.95)}

\item{population}{population size (optional)}

\item{individuals}{title of individuals (default is "individuals")}

\item{question}{"survey question to be asked to individuals" (optional)}

\item{answer}{survey answer choice to be considered a positive response}
}
\value{
a vector with first element as question, second element as correct answer, and other elements as other answer choices.
}
\description{
Make a Confidence Interval for Proportion Question
}
\examples{
makeQA_CIproportion()
}
