% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamesFunctions.R
\name{makeQA_EventProbability}
\alias{makeQA_EventProbability}
\title{Generate probability question}
\usage{
makeQA_EventProbability(type = NULL)
}
\arguments{
\item{type}{Determines the type of probability question.  Valid inputs: 1, 2, 3}
}
\value{
a vector with first element as question, second element as correct answer, and other elements as other answer choices.
}
\description{
Generate probability question
}
\examples{
makeQA_EventProbability(type = 1)
"Events A and B are independent.The probability of A is 0.9.  The probability of B is 0.3.  What is the probability that neither A nor B will occur?"
"0.070"
"0.600"
"0.270"
"1.200"
"0.930"
}
