% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamesFunctions.R
\name{makeQA_ExpectedValue}
\alias{makeQA_ExpectedValue}
\title{Generate expected value function}
\usage{
makeQA_ExpectedValue(prob = NULL, values = NULL)
}
\arguments{
\item{prob}{Vector of probabilities corresponding to values vector.  Must sum to 1 and be non-negative}

\item{values}{Vector of values that r.v. X takes}
}
\value{
a vector with first element as question, second element as correct answer, and other elements as other answer choices.
}
\description{
Generate expected value function
}
\examples{
makeQA_ExpectedValue(prob = c(0.4, 0.2, 0.4), values = c(2,4,8))
"f(2) = 0.4.  f(4) = 0.2.  f(8) = 0.4.  What is the expected value of X?"
"4.8"
"14"
"15"
"84.672"
"5"
}
