% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamesFunctions.R
\name{makeQA_HypothesisTest}
\alias{makeQA_HypothesisTest}
\title{Generate hypothesis test question}
\usage{
makeQA_HypothesisTest(type = NULL, level = 0.05, Xbar = NULL,
  sigma = NULL, n = NULL, mu_0 = NULL)
}
\arguments{
\item{type}{Type of test 0:  Left tail, 1: Right tail, 2: Two tailed}

\item{level}{Level of test.  Must be great than 0 and less than 0.5}

\item{Xbar}{Sample mean}

\item{sigma}{Sample standard deviation}

\item{n}{Sample size}

\item{mu_0}{What mu equals under null hypothesis}
}
\value{
a vector with first element as question, second element as correct answer, and other elements as other answer choices.
}
\description{
Generate hypothesis test question
}
\examples{
makeQA_HypothesisTest(type = 0)
"H_0: mu = 4.4; H_a = mu < 4.4.  You take 43 samples and find that the sample mean is 4.312.  The sample standard deviation is 1.9. Do you reject the null hypothesis at the 0.05 level?"
"yes"
"no"

makeQA_HypothesisTest(type = 1)
"H_0: mu = 7; H_a = mu > 7.  You take 65 samples and find that the sample mean is 7.109.  The sample standard deviation is 4.2. Do you reject the null hypothesis at the 0.05 level?"
"no"
"yes"
}
